chore: BL-15916 separate raw source data files from processed data files
https://issues.bloomlibrary.org/youtrack/issue/BL-15916

The card notes that the raw source files and the files our pipeline generates
from them lived together in one folder and should be separated. Move the raw
source files into a new language-data/source/ subfolder, leaving the generated
files (languageData.json, equivalentTags.json), which the runtime imports, in
language-data/:

- language-data/source/langtags.json
- language-data/source/iso-639-3.tab
- language-data/source/iso-639-3-macrolanguages.tab

Updated the source-file import/read paths in langtagProcessing.ts and
langtagProcessingHelpers.ts, and the Data sources section of the README.

Verified: re-running the langtag-processing pipeline from the new paths
produces byte-identical languageData.json and equivalentTags.json; tsc
typecheck passes; the full find-language test suite passes (174/174).
